Extract findIlmoForDate function to ilmos db-submodule
